requirements.txt
astroid==2.0.4
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.1.2
django-ckeditor==5.6.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-js-asset==1.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
itypes==1.1.0
jedi==0.13.1
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==2.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
pydot==1.2.4
Pygments==2.3.0
PyJWT==1.6.4
pylint==2.1.1
pyparsing==2.3.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
traitlets==4.3.2
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
wrapt==1.10.11
